requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
openai>=0.27.0
langchain>=0.0.200
python-dotenv>=0.19.0
langchain_community
langchain_experimental
tabulate